+++ packages/ricos-orchestrator/web/src/helpers-config.ts
@@ -1,5 +1,5 @@
 import type { RicosEditorProps } from 'ricos-common';
-import { GALLERY_TYPE, IMAGE_TYPE, LINK_PREVIEW_TYPE } from 'ricos-content';
+import { GALLERY_TYPE, IMAGE_TYPE, LINK_PREVIEW_TYPE, POLL_TYPE } from 'ricos-content';
 
 // TODO: move these configs to proper plugins
 export const getHelpersConfig = (props: RicosEditorProps) => ({
@@ -10,6 +10,9 @@ export const getHelpersConfig = (props: RicosEditorProps) => ({
   [GALLERY_TYPE]: {
     handleFileUpload: props._rcProps?.helpers?.handleFileUpload,
     handleFileSelection: props._rcProps?.helpers?.handleFileSelection,
+  },
+  [POLL_TYPE]: {
+    handleFileUpload: props._rcProps?.helpers?.handleFileUpload,
   },
   [LINK_PREVIEW_TYPE]: {
     linkPanelSettings: props.linkPanelSettings,